Separate token logic into its own file and import functions
